requirements.txt
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
google-api-core==1.21.0
google-auth==1.19.1
google-cloud-vision==1.0.0
google-cloud-storage==1.29.0
googleapis-common-protos==1.52.0
grpcio==1.30.0
idna==2.10
joblib==0.16.0
numpy==1.19.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
requests==2.24.0
rsa==4.6
scikit-learn==0.23.1
scipy==1.5.1
six==1.15.0
threadpoolctl==2.1.0
urllib3==1.25.9



